show all product, add button back, add textchange_IP, update productG…

+++ DAL/Product_DAL.cs
@@ -163,7 +163,7 @@ namespace DAL
                 Unit_P = i["Unit_P"].ToString(),
                 Price_P = i["Price_P"].ToString(),
                 VAT = i["VAT"].ToString(),
-                VATInclusive_P = Convert.ToInt32(i["VAT_Inclusive_P"].ToString()),
+                VATInclusive_P = Convert.ToDouble(i["VAT_Inclusive_P"].ToString()),
                 IMG_P = (byte[])(i["IMG_P"]),
                 Status = Convert.ToBoolean(i["Status"].ToString())
             };
